+++ app/src/main/java/com/example/leonidsnotesapplication/presentation/notes_feature/DeleteDialogFragment.kt
@@ -1,4 +1,4 @@
-package com.example.leonidsnotesapplication.presentation.notes_feature.util
+package com.example.leonidsnotesapplication.presentation.notes_feature
 
 import android.app.AlertDialog
 import android.app.Dialog
