+++ app/src/main/res/layout/create_folder_dialog.xml
@@ -31,6 +31,7 @@
             app:layout_constraintEnd_toEndOf="parent"/>
 
         <Button
+            android:id="@+id/cancelButton"
             android:layout_width="80dp"
             android:layout_height="30dp"
             android:layout_marginStart="24dp"
@@ -43,6 +44,7 @@
             app:layout_constraintTop_toBottomOf="@id/etTitleInput"/>
 
         <Button
+            android:id="@+id/createButton"
             android:layout_width="80dp"
             android:layout_height="30dp"
             android:layout_marginTop="20dp"
